Add implementation for CGRA task mapping and shared placement utilities
