+++ lib/TaskflowDialect/Util/CMakeLists.txt
@@ -1,0 +1,15 @@
+add_mlir_library(MLIRTaskflowUtil
+    CgraPlacementUtils.cpp
+    AllocateCgraTaskMapper.cpp
+
+    DEPENDS
+    MLIRTaskflowTransformsIncGen
+
+    LINK_LIBS PUBLIC
+    MLIRIR
+    MLIRPass
+    MLIRSupport
+    MLIRFuncDialect
+    MLIRTaskflow
+    LLVMSupport
+)
